requirements.txt
discord
python-dotenv
dotenv
mongo
pymongo
pillow
aiohttp